created agent state class, greeter class, and fixed up agent class
