+++ WhatsCooking.py
@@ -14,6 +14,9 @@ from langgraph.checkpoint.sqlite import SqliteSaver
 from tavily import TavilyClient
 from urllib.parse import urlparse
 
+from agent_state import AgentState
+from greeter import GreeterAgent
+
 # initializes OpenAI model
 model = ChatOpenAI(model="gpt-3.5-turbo", temperature = 0)
 
@@ -39,36 +42,36 @@ class Dishes(BaseModel):
     dishesList: List[str]
 
 # stores inputs and outputs for nodes
-class AgentState(TypedDict):
-    userDecision: UserDecision
-    preferences: str
-    dishSearchResults: List[str]
-    maxDishSearchResults: int
-    dishesFromSearch: List[str]
-    dishesSeen: List[str]
-    dishesToShow: List[str]
-    domainsVisited: List[str]
-    maxRecommendations: int
-    dishResearchResults: List[str]
-    maxDishResearchResults: int
+# class AgentState(TypedDict):
+#     userDecision: UserDecision
+#     preferences: str
+#     dishSearchResults: List[str]
+#     maxDishSearchResults: int
+#     dishesFromSearch: List[str]
+#     dishesSeen: List[str]
+#     dishesToShow: List[str]
+#     domainsVisited: List[str]
+#     maxRecommendations: int
+#     dishResearchResults: List[str]
+#     maxDishResearchResults: int
 
 # system prompts for agents
-GREETER_PROMPT = """You are a professional recipe recommender inquiring about what kind of recipe the user \
-would like to cook. Make sure to greet the user. You must ask what kind of food they are in the mood for. Tell \
-the user if they do not know what they are in the mood for that is ok. Ask if the user has any other preferences \
-and give some examples of types of preferences. Don't say anything after asking for preferences. If the user gives \
-preferences, make sure they are food related. If the preferences are food related or they have no preferences, respond \
-with the following output:
+# GREETER_PROMPT = """You are a professional recipe recommender inquiring about what kind of recipe the user \
+# would like to cook. Make sure to greet the user. You must ask what kind of food they are in the mood for. Tell \
+# the user if they do not know what they are in the mood for that is ok. Ask if the user has any other preferences \
+# and give some examples of types of preferences. Don't say anything after asking for preferences. If the user gives \
+# preferences, make sure they are food related. If the preferences are food related or they have no preferences, respond \
+# with the following output:
 
-{'decision': "valid",
-'preferences': <user preferences>}
+# {'decision': "valid",
+# 'preferences': <user preferences>}
 
-If the preferences are not food related, tell the person sorry and kindly say you can only accept food related preferences. \
-Respond with the following output:
+# If the preferences are not food related, tell the person sorry and kindly say you can only accept food related preferences. \
+# Respond with the following output:
 
-{'decision': "insufficientResponse",
-'clarifyingRespone': <message to user>}
-"""
+# {'decision': "insufficientResponse",
+# 'clarifyingRespone': <message to user>}
+# """
 
 DISH_SEARCH_PROMPT = """You are a researcher with the task of finding food recipes. You may be given preferences \
 about the types of food recipes to find. Generate a list of search queries to find relevant food recipes. Only generate \
@@ -172,11 +175,13 @@ def question_user(questionToUser: str, systemPrompt: str) -> UserDecision:
         userDecision = model.with_structured_output(UserDecision).invoke(messages)
     return userDecision
 
-# greeter agent
-def greeter_node(state: AgentState):
+greeter_agent = GreeterAgent()
 
-    userDecision = question_user("Hello! What kind of food are you in the mood for today? If you're not sure, that's totally okay. Do you have any preferences such as cuisine type (Italian, Mexican, Asian), dietary restrictions (vegetarian, gluten-free), or specific ingredients you'd like to include?", GREETER_PROMPT)
-    return {"userDecision": userDecision, "preferences": userDecision.preferences}
+# # greeter agent
+# def greeter_node(state: AgentState):
+
+#     userDecision = question_user("Hello! What kind of food are you in the mood for today? If you're not sure, that's totally okay. Do you have any preferences such as cuisine type (Italian, Mexican, Asian), dietary restrictions (vegetarian, gluten-free), or specific ingredients you'd like to include?", GREETER_PROMPT)
+#     return {"userDecision": userDecision, "preferences": userDecision.preferences}
 
 # dish searcher agent
 def dish_searcher_node(state: AgentState):
@@ -276,7 +281,7 @@ def check_post_show_dish_decision(state: AgentState):
 builder = StateGraph(AgentState)
 
 # adds nodes to graph
-builder.add_node("greeter", greeter_node)
+builder.add_node("greeter", greeter_agent.run)
 builder.add_node("dish_search", dish_searcher_node)
 builder.add_node("list_former", dish_list_former_node)
 builder.add_node("show_dishes", show_dishes_node)
